requirements.txt
google-cloud-vision==3.7.2
google-cloud-storage==2.14.0

